Fancy new bash programmable completions
I now have a special file for custom bash completions, which for some reason
does not work with my bash in Arch. It gets sourced in the bashrc, but before
the default bash completion script to let the defaults override them if they
already exist. The completions for evince and pdflatex should make writing LaTeX
much less irritating.

The 'open' alias in bash_aliases is now a function. This allows us to
automatically redirect output to /dev/null, since most of the time we're opening
some sort of media.