Update git branch in PS1
Rather than stripping out all of the whitespace, only trim the first two
characters (which is the asterisk and the leading space).

Also eliminate any parentheses, which might be present during a rebase or in a
detached HEAD state.

The argument to "[ -z" must be in quotes in the event that the string contains
spaces.